fix(tvix/eval): PoC fix for b/261 by finalising only on demand
We track whether we need to finalise `idx` by storing a boolean in the
phantom `needs_finalise_idx`. This is of course quite ugly. If we want
to continue on this road, we should at least add an internal sentinel
value for this purpose.

Code needs to be cleaned up, can we avoid having an extra local on the
stack for every default expression?

Change-Id: Ic04fb80ec671a2ba11fa645090769c335fb7f58b