feat(tvix/eval): impl gc::Trace for NixAttrs
This required manual implementations on all the relevant types, for
different reasons.

On AttrsRep, the contained `imbl::OrdMap` type doesn't allow use of
the macro as it comes from a third-party crate.

On NixAttrs, the macros can not be used because they generate an empty
`Drop` implementation. They do this because they *don't* want people
to have `Drop` implementations that do something and are trying to
enforce that statically, however, this does not actually work because
we can not have a `Drop` implementation *at all* (as we are partially
moving out of this type frequently).

Change-Id: Icd91ca8eb7b012aa3e7f5f06b653412faae3fe6b