feat(tvix/nix-compat): supports serialization of BString as UTF-8 strings
Previously, we were serializing BString as a &[u8], i.e. a list of u8.

This makes sense when you consider that BString are *MOSTLY* valid UTF-8 and not
guaranteed valid UTF-8 like String.

At (de)serialization boundaries, we are not too much interested by those concerns though,
we will brutally assume this has to be valid UTF-8 and move on.

This makes (de)serialization of Derivation structure via serde, for, e.g. JSON, possible.

In the future, we will need to accommodate for UTF-16 shenigans provided
by Windows constraints, e.g. perform the adequate escaping of UTF-16 in UTF-8.

This is an open question though and one we should not chase too much.

Change-Id: I5afd03eb4b261751f8df79b3d4dd87244ea461fc