feat(tazjin/gio-list-apps): CLI to list xdg apps using Gio library
This does nothing other than dump out the list of XDG apps to stdout
in JSON format. There are no options or anything.

This can be used for selection in app launchers (e.g. dmenu, something
based on completing-read in emacs, rofi, etc.).

I wrote this because I don't want to deal with having to do this in
Elisp. It's also unclear what logic actually hides behind under the
hood here, so why not just use the official library.

Change-Id: I16fed2c92760cadecc02c59a4e537a1fa247aff9