feat(tvix/store): fix ctrl-c handling on mount command
This registers moves the run-and-wait-until-finished of the filesystem
to a separate task.

We install a signal handler in another task, that's able to unmount the
filesystem as well.

The "main thread" waits on both tasks to complete.

This avoids having disconnected mountpoints when Ctrl-C'ing a
`tvix-store mount` invocation.

It also still successfully exits if the filesystem is externally
unmounted.

Change-Id: I149f705a6cb50188177f2a6c6a5fcd77218e2a3f