fix(tvix/eval): fix b/281 by adding Value::Catchable
This commit makes catchable errors a variant of Value.

There is quite a lot of boilerplate in this commit.

Mostly done, and worth reviewing, but still at least two things left
for me to do (plus polishing):

- There is one FIXME where I didn't know how to synthesize a span
  correctly; tazjin@ can you advise on the correct way to do that?

- Please see also the comment for the variant
  ErrorKind::CatchableErrorKind; this variant should be eliminated (it
  is a footgun).  The CL to eliminate it is straightforward but very
  noisy, and this CL is already large enough as it is.

The main downside of this approach is that we lose the ability to
use Rust's `?` syntax for propagating catchable errors.  This isn't
needed as often as you might think; the only place it is annoying
seems to be in the string concatenation and equality-testing routines.

Change-Id: Ibe89438d8a70dcec29e016df692b5bf88a5cad13