fix(tvix/store/grpc): Use block_in_place to avoid nested async runtimes
The BlobService trait is sync Rust, but the gRPC client is async so
we're using `block_on` to call async code from sync code. However, the
daemon starts in an async runtime, so we end up calling `block_on` while
inside an async runtime which causes a panic.

As a quick workaround, we can use `block_in_place` which will force any
async tasks on the thread to move to a different thread which makes it
safe to spawn a new async runtime on the current thread with `block_on`.

A better solution would be to make the `BlobService` trait async so we
don't have to go from async->sync->async.

Change-Id: I7bf2769b413520bf1e17ceb3f0982f7a61f802ce