fix(tvix/store/grpc): Use spawn_blocking when calling async->sync
In the gRPC server implementations, we are calling potentially long
blocking sync code from async code. The proper way to do this is to use
a `spawn_blocking` call which will prevent the sync call from locking up
the async runtime.

This happens to also fix the issue where we were having nested async
runtimes, as it is okay to call `block_on` inside a `spawn_blocking`
since the `spawn_blocking` closure is executed in a non-async context.

Issue: https://b.tvl.fyi/issues/304
Change-Id: I4830780d3012207e7628e34610a98bbee66a6e16