test(tvix/store): Use rstest to test all service backends
Previously, we were only testing the memory implementation of our
services. This lead to a bug in the gRPC implementations of the services
to regress and panic.

This adds `rstest` and some new helpers to make it easy to write tests
which exercise all of the backends.

These new test cases are able to catch the issue fixed by cl/9294.

Change-Id: I6aed816e41df3256d1ea283002c98c86ebddafd2