feat(tvix/eval): use std::Weak for self-referencing upvalues
This CL contains the result of a largely-unsuccessful experiment,
and performance numbers.

Currently, tvix uses reference-counting for garbage collection.  The
main problem with reference counting is, of course, that it can't
collect cycles.  There are only two things in tvix that can create
cycles:

  1. Rc::cyclic_new() used in the builtins
  2. Thunk, which contains a RefCell

The first cycle is not a problem; the builtins are expected to hang
around for the entire evaluation, so it's okay to leak them.  They
also can't have references to anything else.

This CL attempted to reduce the number of cycles created by thunks
by replacing the reference from a thunked closure's upvales to the
closure itself with an rc::Weak.

I also tried delete values from Upvalues as they are used in order
to allow earlier collection.

The statistics below come from `tvix --no-warnings -E '(import
/git/ubik/nixpkgs {}).chromium.outPath'` wrapped by `time -v`:

Baseline:

        User time (seconds): 117.37
        Maximum resident set size (kbytes): 492260

With this CL:

        User time (seconds): 107.35
        Maximum resident set size (kbytes): 482368

cppnix:

        User time (seconds): 1.43
        Maximum resident set size (kbytes): 269628

So, this really didn't make a difference.

I'm starting to think maybe we shouldn't worry about GC yet.  We're
using less than 2x the amount of memory that cppnix uses, but 100x
the amount of wall-clock time.

Change-Id: I740b1dcf5183f3f32881841e7e7bd2c6523f8f9b