refactor(tvix): move castore into tvix-castore crate
This splits the pure content-addressed layers from tvix-store into a
`castore` crate, and only leaves PathInfo related things, as well as the
CLI entrypoint in the tvix-store crate.

Notable changes: fixtures and utils had to be moved out of the test
module to a general purpose module, so they can be used from tvix-store.

The (now two) golang targets are commented out, as it's not possible to
update them properly in the same CL.

Bug: https://b.tvl.fyi/issues/301

Change-Id: I8d675d4bf1fb697eb7d479747c1b1e3635718107