chore(third_party/josh): bump to latest HEAD
It seems our currently pinned josh commit doesn't seem to support
pushing back, and just timeouts when trying to do so.

Bump to the latest head, let's see how it works out there.

We don't need to pull a more recent rustc, as the nixpkgs rustc seems to
be recent enough.

Change-Id: I4f6d775df4db13a4537049292edfe969d2bb45ea