fix(tools/magrathea): isolate from environment packages
This makes magrathea use repostory version of nix and git. This is
done in the persuit of enforcing guaranteed unified experience of
magrathea tool across any user of `tvl-kit`-based
repositories. Especially among ubuntu users with uncontrolled set of
packages and versions installed on their system.

Not having this fix was giving build problems for `mg build` for one
of the users as he had 2.17 version of nix that had inconsistent hash
computation with 2.3.

Change-Id: I3182faf4c545ac61f6cc1cc862dc23d51c1cd397