experiment: highly unsafe 7% speedup of tvix-eval
With this (highly experimental, not actually suggesting to submit
this) change, tvix-eval only allocates once per generator (to pin the
future) and avoids an additional Rc.

I wanted to see how difficult this would be to do in Rust. Not super
difficult, but not comfortable.

The biggest problem we'd face with making this safer is that we don't
have a clear concept of the lifetime of the VM object, to which a lot
of stuff is tied.

Change-Id: I300f67964ff353bf9a3229b9c346d517169d488c